refactor: share the async retry-query decorator instead of copying it per transport
Sonar's new-code duplication gate went red at 3.3% against a 3% threshold, and
it was right: the three `RetryQueryHandlerDecoratorAsync` classes added in the
last commit were byte-identical apart from the transport name, as are the three
synchronous ones they were copied from.

The logic now lives in `ARetryQueryHandlerDecoratorAsync` in
Transport.RelationalDatabase, with the pipeline name supplied by an abstract
`PolicyName`; each transport keeps a subclass that is a constructor and one
expression. A transient read is the same transient read on every transport, so
there was nothing per-transport in the copies.

Its behaviour is tested once, in RelationalDatabase.Tests, rather than three
times - including that the pipeline the subclass names is the one that runs,
which is the part each transport depends on. That the decorator resolves at all
is what container verification in each transport's own tests covers.

Also drops two duplicate `using` directives in the PostgreSQL test file the
previous commit added (CS0105), a casualty of generating it from a sibling whose
using block was ordered differently.
